# libs/shared/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/platform/users": {
      "get": {
        "operationId": "PlatformController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformUserListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取平台用户列表",
        "tags": [
          "Platform"
        ]
      },
      "post": {
        "operationId": "PlatformController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建平台用户",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/users/{id}/activate": {
      "post": {
        "operationId": "PlatformController_activateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformUserActivationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "启用平台用户",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/users/{id}/deactivate": {
      "post": {
        "operationId": "PlatformController_deactivateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformUserActivationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "停用平台用户",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/users/{id}/roles": {
      "post": {
        "operationId": "PlatformController_assignUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanitizedUserWithOrgUnits"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分配用户角色",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/users/{id}/org-memberships": {
      "post": {
        "operationId": "PlatformController_assignUserOrgMemberships",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserOrgMembershipsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanitizedUserWithOrgUnits"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分配用户组织关系",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/roles": {
      "get": {
        "operationId": "PlatformController_listRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRoleListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取平台角色列表（基础查询）",
        "tags": [
          "Platform"
        ]
      },
      "post": {
        "operationId": "PlatformController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRoleSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建平台角色",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/roles/{id}/permissions": {
      "post": {
        "operationId": "PlatformController_assignRolePermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolePermissionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformRoleSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分配角色权限（全量替换）",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/org-units": {
      "get": {
        "operationId": "PlatformController_listOrgUnits",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取平台组织列表（基础查询）",
        "tags": [
          "Platform"
        ]
      },
      "post": {
        "operationId": "PlatformController_createOrgUnit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgUnitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建组织单元",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/org-units/tree": {
      "get": {
        "operationId": "PlatformController_listOrgUnitTree",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitTreeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取平台组织树",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/org-units/{id}": {
      "get": {
        "operationId": "PlatformController_getOrgUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取组织单元详情",
        "tags": [
          "Platform"
        ]
      },
      "patch": {
        "operationId": "PlatformController_updateOrgUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrgUnitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新组织单元基本信息",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/org-units/{id}/activate": {
      "post": {
        "operationId": "PlatformController_activateOrgUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrgUnitActivationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "启用组织单元",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/org-units/{id}/deactivate": {
      "post": {
        "operationId": "PlatformController_deactivateOrgUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrgUnitActivationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "停用组织单元",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/org-units/{id}/move": {
      "post": {
        "operationId": "PlatformController_moveOrgUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveOrgUnitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformOrgUnitSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "移动组织单元并调整排序",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/navigation": {
      "get": {
        "operationId": "PlatformController_getAllNavigationItems",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavigationListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取完整导航树（平台管理员只读视图）",
        "tags": [
          "Platform"
        ]
      }
    },
    "/api/platform/navigation/sync": {
      "post": {
        "operationId": "PlatformController_syncNavigation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavigationSyncSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "记录当前导航 SSOT 的同步审计快照",
        "tags": [
          "Platform"
        ]
      }
    },
...
  },
  "info": {
    "title": "POMS API",
    "description": "Project Oriented Management System API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SanitizedUserWithOrgUnits": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "displayName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "platform:users:manage",
                "platform:roles:manage",
                "platform:navigation:manage",
                "platform:org-units:manage",
                "commission:rule-versions:manage",
                "commission:assignments:manage",
                "commission:calculations:manage",
                "commission:payouts:manage",
                "commission:adjustments:manage",
                "contract:finance:manage",
                "project:read",
                "project:write",
                "project:delete",
                "nav:dashboard:view",
                "nav:platform:view",
                "nav:projects:view",
                "nav:contracts:view",
                "nav:profile:view"
              ]
            }
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "nullable": true
          },
          "emailVerified": {
            "type": "boolean"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "orgUnits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitOrg"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "username",
          "roles",
          "permissions",
          "email",
          "avatarUrl",
          "isActive",
          "lastLoginAt",
          "emailVerified",
          "phoneVerified",
          "phone",
          "orgUnits"
        ],
        "id": "SanitizedUserWithOrgUnits"
      },
...
      "PlatformUserListDto": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PlatformUserSummary"
        },
        "id": "PlatformUserList"
      },
      "CreatePlatformUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "primaryOrgUnitId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "nullable": true
          },
          "initialRoleIds": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": [
          "username",
          "displayName",
          "primaryOrgUnitId"
        ],
        "id": "CreatePlatformUserRequest"
      },
      "UpdatePlatformUserActivationRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          },
          "comment": {
            "type": "string",
            "maxLength": 1000
          },
          "expectedVersion": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "id": "UpdatePlatformUserActivationRequest"
      },
      "AssignUserRolesRequest": {
        "type": "object",
        "properties": {
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          "reason": {
            "type": "string",
            "maxLength": 1000
          },
          "expectedVersion": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "roleIds"
        ],
        "id": "AssignUserRolesRequest"
      },
      "AssignUserOrgMembershipsRequest": {
        "type": "object",
        "properties": {
          "primaryOrgUnitId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "secondaryOrgUnitIds": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          "reason": {
            "type": "string",
            "maxLength": 1000
          },
          "expectedVersion": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "primaryOrgUnitId"
        ],
        "id": "AssignUserOrgMembershipsRequest"
      },
...
      "PlatformRoleListDto": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PlatformRoleSummary"
        },
        "id": "PlatformRoleList"
      },
      "CreateRoleRequest": {
        "type": "object",
        "properties": {
          "roleKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "roleKey",
          "name"
        ],
        "id": "CreateRoleRequest"
      },
      "AssignRolePermissionsRequest": {
        "type": "object",
        "properties": {
          "permissionKeys": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "platform:users:manage",
                "platform:roles:manage",
                "platform:navigation:manage",
                "platform:org-units:manage",
                "commission:rule-versions:manage",
                "commission:assignments:manage",
                "commission:calculations:manage",
                "commission:payouts:manage",
                "commission:adjustments:manage",
                "contract:finance:manage",
                "project:read",
                "project:write",
                "project:delete",
                "nav:dashboard:view",
                "nav:platform:view",
                "nav:projects:view",
                "nav:contracts:view",
                "nav:profile:view"
              ]
            }
          }
        },
        "required": [
          "permissionKeys"
        ],
        "id": "AssignRolePermissionsRequest"
      },
...
      "PlatformOrgUnitListDto": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PlatformOrgUnitSummary"
        },
        "id": "PlatformOrgUnitList"
      },
...
      "PlatformOrgUnitTreeDto": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PlatformOrgUnitTreeDto__schema0"
        },
        "id": "PlatformOrgUnitTree"
      },
...
    }
  }
}